backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.5.3
flask-limiter==3.5.0
mongoengine==0.27.0
pymongo==4.6.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
reportlab==4.0.7
Werkzeug==3.0.1

